chore(mc): Put TopSitesForm, Editor into their own files (nobug)
