+++ system-addon/content-src/components/TopSites/TopSiteForm.jsx
@@ -1,0 +1,161 @@
+const React = require("react");
+const {actionCreators: ac, actionTypes: at} = require("common/Actions.jsm");
+const {FormattedMessage} = require("react-intl");
+
+const TOP_SITES_SOURCE = "TOP_SITES";
+
+class TopSiteForm extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      label: props.label || "",
+      url: props.url || "",
+      validationError: false
+    };
+    this.onLabelChange = this.onLabelChange.bind(this);
+    this.onUrlChange = this.onUrlChange.bind(this);
+    this.onCancelButtonClick = this.onCancelButtonClick.bind(this);
+    this.onAddButtonClick = this.onAddButtonClick.bind(this);
+    this.onSaveButtonClick = this.onSaveButtonClick.bind(this);
+    this.onUrlInputMount = this.onUrlInputMount.bind(this);
+  }
+  onLabelChange(event) {
+    this.resetValidation();
+    this.setState({"label": event.target.value});
+  }
+  onUrlChange(event) {
+    this.resetValidation();
+    this.setState({"url": event.target.value});
+  }
+  onCancelButtonClick(ev) {
+    ev.preventDefault();
+    this.props.onClose();
+  }
+  onAddButtonClick(ev) {
+    ev.preventDefault();
+    if (this.validateForm()) {
+      let site = {url: this.cleanUrl()};
+      if (this.state.label !== "") {
+        site.label = this.state.label;
+      }
+      this.props.dispatch(ac.SendToMain({
+        type: at.TOP_SITES_ADD,
+        data: {site}
+      }));
+      this.props.dispatch(ac.UserEvent({
+        source: TOP_SITES_SOURCE,
+        event: "TOP_SITES_ADD"
+      }));
+      this.props.onClose();
+    }
+  }
+  onSaveButtonClick(ev) {
+    ev.preventDefault();
+    if (this.validateForm()) {
+      let site = {url: this.cleanUrl()};
+      if (this.state.label !== "") {
+        site.label = this.state.label;
+      }
+      this.props.dispatch(ac.SendToMain({
+        type: at.TOP_SITES_PIN,
+        data: {site, index: this.props.index}
+      }));
+      this.props.dispatch(ac.UserEvent({
+        source: TOP_SITES_SOURCE,
+        event: "TOP_SITES_EDIT",
+        action_position: this.props.index
+      }));
+      this.props.onClose();
+    }
+  }
+  cleanUrl() {
+    let url = this.state.url;
+    // If we are missing a protocol, prepend http://
+    if (!url.startsWith("http:") && !url.startsWith("https:")) {
+      url = `http://${url}`;
+    }
+    return url;
+  }
+  resetValidation() {
+    if (this.state.validationError) {
+      this.setState({validationError: false});
+    }
+  }
+  validateUrl() {
+    try {
+      return !!new URL(this.cleanUrl());
+    } catch (e) {
+      return false;
+    }
+  }
+  validateForm() {
+    this.resetValidation();
+    // Only the URL is required and must be valid.
+    if (!this.state.url || !this.validateUrl()) {
+      this.setState({validationError: true});
+      this.inputUrl.focus();
+      return false;
+    }
+    return true;
+  }
+  onUrlInputMount(input) {
+    this.inputUrl = input;
+  }
+  render() {
+    return (
+      <form className="topsite-form">
+        <section className="edit-topsites-inner-wrapper">
+          <div className="form-wrapper">
+            <h3 className="section-title">
+              <FormattedMessage id={this.props.editMode ? "topsites_form_edit_header" : "topsites_form_add_header"} />
+            </h3>
+            <div className="field title">
+              <input
+                type="text"
+                value={this.state.label}
+                onChange={this.onLabelChange}
+                placeholder={this.props.intl.formatMessage({id: "topsites_form_title_placeholder"})} />
+            </div>
+            <div className={`field url${this.state.validationError ? " invalid" : ""}`}>
+              <input
+                type="text"
+                ref={this.onUrlInputMount}
+                value={this.state.url}
+                onChange={this.onUrlChange}
+                placeholder={this.props.intl.formatMessage({id: "topsites_form_url_placeholder"})} />
+              {this.state.validationError &&
+                <aside className="error-tooltip">
+                  <FormattedMessage id="topsites_form_url_validation" />
+                </aside>
+              }
+            </div>
+          </div>
+        </section>
+        <section className="actions">
+          <button className="cancel" type="button" onClick={this.onCancelButtonClick}>
+            <FormattedMessage id="topsites_form_cancel_button" />
+          </button>
+          {this.props.editMode &&
+            <button className="done save" type="submit" onClick={this.onSaveButtonClick}>
+              <FormattedMessage id="topsites_form_save_button" />
+            </button>
+          }
+          {!this.props.editMode &&
+            <button className="done add" type="submit" onClick={this.onAddButtonClick}>
+              <FormattedMessage id="topsites_form_add_button" />
+            </button>
+          }
+        </section>
+      </form>
+    );
+  }
+}
+
+TopSiteForm.defaultProps = {
+  label: "",
+  url: "",
+  index: 0,
+  editMode: false // by default we are in "Add New Top Site" mode
+};
+
+module.exports = TopSiteForm;
